+++ src/lib/view.mjs
@@ -1,6 +1,3 @@
-
-import os from "os";
-import { EOL } from "os";
 import { stdout } from "process";
 
 
@@ -16,30 +13,6 @@ function truncateString(str, limit) {
     }
 }
 
-function sortFiles(list) {
-    list.sort(function (a, b) {
-        if (a.Type > b.Type) return 1;
-        if (a.Type < b.Type) return -1;
-
-        if (a.name > b.name) return 1;
-        if (a.name < b.name) return -1;
-    })
-};
-
-function getEOLSymbol() {
-  return EOL.replace(/\n/g,'\\n').replace(/\t/,'\\t');
-};
-
-function showCPUS() {
-  const cpus = os.cpus();
-  printGreenText(`CPUS: ${cpus.length}${EOL}`);
-
-  cpus.forEach((cpu) => {
-    const text = `model: ${cpu.model.trim()}, clock rate: ${(cpu.speed / 1024).toFixed(4)} GHz${EOL}`;
-    printGreenText(text);
-  });
-}
-
 function printGreenText(text) {
   stdout.write(colorize(92, text));
 }
@@ -48,5 +21,4 @@ function printRedText(text) {
   stdout.write(colorize(91, text));
 }
 
-export { colorize, truncateString, sortFiles, getEOLSymbol, 
-  showCPUS, printGreenText, printRedText };
+export { colorize, truncateString, printGreenText, printRedText };
